Add functionality to disassociate IP address from the server, before deleting the server.

+++ lib/hpcloud/commands/servers/remove.rb
@@ -17,14 +17,18 @@ Aliases: servers:delete, servers:del
         begin
           # setup connection for compute service
           compute_connection = connection(:compute)
-          #server = compute_connection.servers.get(id)
-          compute_connection.servers.filters = {'instance-id' => ["#{id}"]}
-          server = compute_connection.servers.first
+          server = compute_connection.servers.select {|s| s.id == id}.first
         rescue Excon::Errors::Forbidden => error
           display_error_message(error, :permission_denied)
         end
-        if server
+        if (server && server.id == id)
           begin
+            # disassociate server from address, and release address
+            server.addresses.each do |addr|
+              addr.server = nil
+              addr.destroy
+            end
+            # now delete the server
             server.destroy
             display "Removed server '#{id}'."
           rescue Excon::Errors::Conflict, Excon::Errors::Forbidden => error
